XDZ 1、评论系统API完成 2、支付系统新增“获取评论内容”API。     新增评论、删除评论同步更新book.rating字段 3、拦截器放行获取书籍评论、获取评论内容

+++ src/main/java/one/wcy/ebookloaningtool/xdz/pojo/AddReviewRequest.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class AddReviewRequest {
     private String bookId;
-    private int rating;
+    private BigDecimal rating;
     private String comment;
 }
